# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FundFlow REST API
  version: "1.0.0"
  description: |
    Multi-tenant VC/PE fund management platform. Covers accounts, organizations,
    SPVs, entities, members, invites, portfolio data, and authentication.

    ## Authentication
    Two methods are accepted on every `/api/*` endpoint:

    1. **Session cookie** — set by `POST /api/auth/login`. Used by the web UI.
    2. **Bearer token** — `Authorization: Bearer ff_<48 hex chars>`. Generated
       via `POST /api/auth/tokens` (admin-only). Tokens inherit the owner's
       account roles, so an admin token can hit admin-only routes; a non-admin
       token only sees what its owner is allowed to see.

    Token CRUD itself (`/api/auth/tokens` routes) and `/api/auth/change-password`
    require a real session cookie — never a bearer token — so a leaked token
    cannot mint more tokens or lock the user out of their account.

    ## Permissions model
    - **admin** role: full access. Can create/delete organizations, delete any
      account/SPV/entity, and manage API tokens.
    - **non-admin**: scoped to data they participate in (org member/organizer,
      SPV member, entity manager/owner).
    - Permission checks happen identically for cookie- and token-authenticated
      requests.

  contact:
    name: FundFlow
servers:
  - url: /
    description: Same-origin (web app and API share a host)
...
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /api/currencies:
    get:
      tags: [Currencies]
      summary: List the currency catalog
      description: |
        Returns every known currency (active and inactive). `active`
        currencies are the ones selectable as an SPV's currency on creation
        and offered in the dashboard's "view in" dropdown.
      responses:
        '200':
          description: Currency catalog
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Currency' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/currencies/{code}:
    patch:
      tags: [Currencies]
      summary: Activate or deactivate a currency (admin-only)
      description: |
        USD is the platform's base reporting currency and the exchange-rate
        pivot — it can never be deactivated. Activating a currency triggers
        an immediate best-effort exchange-rate fetch rather than waiting for
        the next scheduled refresh.
      parameters:
        - { name: code, in: path, required: true, schema: { type: string }, example: EUR }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [active]
              properties:
                active: { type: boolean }
      responses:
        '200':
          description: Updated currency
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Currency' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/exchange-rates:
    get:
      tags: [Currencies]
      summary: Latest known exchange rates
      description: |
        Rates are expressed as "USD per 1 unit of the currency" (USD itself
        is always 1) and refreshed at most once a day from a public FX feed.
        No history is kept — only the latest rate per currency.
      responses:
        '200':
          description: Exchange rates
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExchangeRates' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/exchange-rates/refresh:
    post:
      tags: [Currencies]
      summary: Force an immediate exchange-rate refresh (admin-only)
      description: Mainly for troubleshooting a stale or missing rate; the daily refresh otherwise runs automatically.
      responses:
        '200':
          description: Refreshed exchange rates
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExchangeRates' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ──── Portfolio ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Set automatically by `POST /api/auth/login`.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ff_<48 hex chars>
      description: |
        Personal API token created via `POST /api/auth/tokens` (admin-only).
        Tokens inherit the owner's roles.
  responses:
    BadRequest:
      description: Validation error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Authentication required or rejected
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Authenticated but lacks permission
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      properties:
        message: { type: string }
      required: [message]
...
    Currency:
      type: object
      properties:
        code: { type: string, example: "EUR", description: "ISO 4217 code." }
        name: { type: string, example: "Euro" }
        symbol: { type: string, example: "€" }
        active: { type: boolean, description: "Whether this currency is selectable for new SPVs and in the dashboard's currency dropdown. USD is always active." }
    ExchangeRates:
      type: object
      properties:
        base: { type: string, example: "USD", description: "Pivot currency — always USD." }
        rates:
          type: object
          additionalProperties: { type: number }
          example: { USD: 1, EUR: 1.0825 }
          description: "USD per 1 unit of each active currency's code (USD itself is always 1)."
        updatedAt: { type: string, format: date-time, nullable: true, description: "Most recent refresh timestamp across all stored rates; null if none have been fetched yet." }
...